pyproject.toml
[tool.poetry]
name = "flask-books"
version = "0.1.0"
description = "Aplikasi Buku API"
authors = ["perymerdeka <tkjasikk@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_books"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-restx = "^1.1.0"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.3"
mysqlclient = "^2.1.1"
flask-migrate = "^4.0.4"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
